requirements.txt
Flask==1.1.2
numpy==1.19.1
pandas==1.1.1
Pillow==7.2.0
requests==2.25.1
scikit-learn==0.24.2
